issue #1245: moved FROM and FROM NAMED clause from OuterQuery to Query so they appear after SELECT updated javadoc
